Added support for handling user/group being specified in the manifest
This adds the command line options for --user and --group to the stage2
handler. It will use this to switch to the specified user/group before execing
the stage3 command.

The values can be either the uid/gid or the user/group name. It will first
attempt to convert it to a number, and if the whole string is numeric, then it
will use that, else it will look it up. The look up is after the pivot_root,
so it will be based on the local filesystem when pivot/chroot is used.

If the username lookup fails, then it will simply continue and execute as
root. This behavior may be something to review and change to have it propagate
an error.