Refactor of init/bootstrap to handle init container handling.
Now supports spawning containers on startup. This changed the old kurma-init
to be kurma-server, and not do any bootstrap handling.

The new kurma-init uses the init package, which handles bootstrapping and
setting up the whole process, including the container manager so it can
initialize containers, and then starting the RPC server using the container
manager.

This also changes the init/bootstrap to use a private runner object. This is
mostly to avoid using package globals.