Reintroduced pivot_root usage. Added privileged host isolator.
Moved from chroot back to using pivot_root, after addressing the issue with
the rootfs from the initrd.

Also instrumented a custom isolator and new flag on stage2 for a privileged
mode. With this, the container will be setup with devtmpfs and get all devices
rather than simply getting selected devices. Additionally, after the
pivot_root, the old root will not be unmounted, it will be left read-only at
/host.