Refactor stage2 launching.
This adds a new stage2/client package which encapsulates the command line
argument handling for the stage2 process.

This moves the stage1/container to use it.

The "privileged" isolator is renamed to "host/privileged".

Adds SIGCHLD handling to init. Also has the stage2 focus use detached mode.
This is partially to prepare for "enter" functionality, since stage2 would
stay alive, be in the task groups, and if its PID was used to look up
namespaces to join, the stage2 PID namespace isn't in the container. Not
detaching was simply done because of no SIGCHLD handling.