Update "enter" tty handling both server side and locally
Ensure the tty configured in rpc is properly assigned to the stage2 process.

Removed non-functional tty handling from the container creation. May revisit
how to make it better in the future.

Locally moved from termbox to termios to simply set stdin to rawmode. This is
working better. It turns of local echo and buffering, but doesn't clear the
whole screen and lose the cursor like termbox did.