Moved our existing init process in and updated the server to invoke it.
The init process sits at the top of the container and manages the processes
underneath it. The server has a unix socket connection to it with a very
simplistic RPC to ask the init to start a process.

The init is pure C, still built in Go and bundled in the one binary. But this
ensures it has a small memory footprint, does very little, and is just a
single threaded process.

Some of the benefits of this apprach are:

Closer to the container SIGCHLD handling.

Able to preserve the container state, even if the app processes exit. This
allows later logging in even if the process died and the container is in the
EXIT state.

More easily able to instrument pre and post start commands. With the pre
command, if it was to setup mounts or what not, without the init process those
would be lost between the pre command and the actual app, since once it exits
that mount namespace is torn down.

More detailed process information and between server restarts persistence.