Cleanup and refactoring around bootstrap.
This revises a lot of logging and error handling in the bootstrap. Ensure that
errors in bootstrap don't completely halt everything, such as just if DHCP
failed or something.

Refactored some service containers to be explicit in configuration, such as
NTP and the console. That way the specific ones have configuration specific to
them, and InitContainers is meant more for user specified containers and no
runtime customization to them. This is mostly around like password/ssh keys
for the console, ntp servers for ntp, etc.

Some cleanup around container management as well.