Implement remote API functionality to broker requests to the local API
This allows an initial container to be specified that can run an application
to proxy the localhost Kurma API to a remote API, but allow
enforcement/validation handling for things like not allowing a remote request
to launch a priviledged container.

This helps to emphasize the power of the local API not making assumptions
about how the initial containers may expose an API that is larger than itself.