Update UID/GID handling for appc 0.6.0
Updated the UID/GID handling in stage2/3 to be inline with the specification
in 0.6.0. It should first look up the UID/GID in /etc/{passwd,group}, and then
attempt to parse it to an integer. This is more consistent with existing unix
handling, rather than going straight to a UID/GID if the string is all
numbers.